requirements.txt
requests
openai
schedule
python-dotenv
feedparser
beautifulsoup4
ipykernel
pandas
matplotlib
duckdb
streamlit
sqlalchemy